backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
python-dotenv==1.0.0
tenacity==8.2.3
python-multipart==0.0.6
google-genai==0.3.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.38.1
